+++ app/services/task.py
@@ -3,8 +3,13 @@ from sqlalchemy.orm import Session
 from app.repositories.task import TaskRepository
 from app.schemas.task import TaskCreate, TaskRead, TaskUpdate
 
+
+class TaskNotFoundError(Exception):
+    pass
+
+
 class TaskService:
-    '''Ключевые операции с задачами, включая бизнес-логику, валидацию и прочее'''
+    """Ключевые операции с задачами, включая бизнес-логику, валидацию и прочее"""
 
     def __init__(self, db: Session):
         self.db = db
@@ -21,6 +26,8 @@ class TaskService:
 
     def update_task(self, task_id: str, payload: TaskUpdate) -> TaskRead:
         task = self.repository.get_by_id(task_id)
+        if task is None:
+            raise TaskNotFoundError
         if payload.title is not None:
             task.title = payload.title
         if payload.completed is not None:
@@ -30,8 +37,7 @@ class TaskService:
 
     def delete_task(self, task_id: str) -> None:
         task = self.repository.get_by_id(task_id)
+        if task is None:
+            raise TaskNotFoundError
         self.repository.delete(task)
         self.db.commit()
-
-class TaskNotFoundError(Exception):
-    pass
